Added validation to input vars
Added validation for `name` input var.  Since it's the common name for all resources we have to make sure it enforces proper validation.  For example: Resource Groups can have "-" in the name, but storage accounts are limited to lowercase and numbers only.  In that case we'll enforce the most restrictive rule to guarantee resource creation.  

The change adds a validation step to the input parameter that checks name is only lower alpha and numeric.